requirements.txt
tweepy==3.3.0
Flask==0.10.1
Flask-SQLAlchemy==1.0
